recogniser/requirements.txt
fastapi [standard] >= 0.115.5
python-telegram-bot>=21.7
transformers==4.38.0
accelerate==0.27.0
soundfile==0.12.1
librosa==0.10.1
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0

